+++ source/rafcon/core/state_machine_manager.py
@@ -48,10 +48,10 @@ class StateMachineManager(Observable):
                 self.add_state_machine(state_machine)
 
     def delete_all_state_machines(self):
-        self.active_state_machine_id = None
         sm_ids = [sm_id for sm_id in self.state_machines]
         for sm_id in sm_ids:
-            self.remove_state_machine(sm_id)
+            if not (sm_id == self.active_state_machine_id):
+                self.remove_state_machine(sm_id)
 
     def open_state_machines(self, state_machine_path_by_sm_id):
         from rafcon.core.storage import storage
@@ -121,13 +121,6 @@ class StateMachineManager(Observable):
             logger.error("There is no state_machine with state_machine_id: %s" % state_machine_id)
             return removed_state_machine
 
-        # a not stopped or finished state machine will stay the active state machine TODO test this and rethink it
-        if state_machine_id is self.active_state_machine_id and \
-                core_singletons.state_machine_execution_engine.finished_or_stopped():
-            if len(self._state_machines) > 0:
-                self.active_state_machine_id = self._state_machines[list(self._state_machines.keys())[0]].state_machine_id
-            else:
-                self.active_state_machine_id = None
         # destroy execution history
         removed_state_machine.destroy_execution_histories()
         return removed_state_machine