+++ source/rafcon/gui/controllers/execution_history.py
@@ -347,11 +347,10 @@ class ExecutionHistoryTreeController(ExtendedController):
             if self.parent is not None and hasattr(self.parent, "focus_notebook_page_of_controller"):
                 # request focus -> which has not have to be satisfied
                 self.parent.focus_notebook_page_of_controller(self)
-                self.model.selected_state_machine_id = self.model.state_machine_manager.active_state_machine_id
 
         if state_machine_execution_engine.status.execution_mode is not StateMachineExecutionStatus.STARTED:
             if not self.model.selected_state_machine_id == self.model.state_machine_manager.active_state_machine_id:
-                self.model.selected_state_machine_id = self.model.state_machine_manager.active_state_machine_id
+                pass
             else:
                 self.update()
 
@@ -363,7 +362,9 @@ class ExecutionHistoryTreeController(ExtendedController):
         self.history_tree_store.clear()
         selected_sm_m = self.model.get_selected_state_machine_model()
         if selected_sm_m:
-            selected_sm_m.state_machine.clear_execution_histories()
+            # this must not be done here!
+            # the core may continue running without the GUI and for this it needs its execution histories
+            # selected_sm_m.state_machine.clear_execution_histories()
             self.update()
 
     def reload_history(self, widget, event=None):