+++ source/rafcon/gui/start.py
@@ -379,8 +379,12 @@ def main():
 
     if core_singletons.state_machine_execution_engine.status.execution_mode == StateMachineExecutionStatus.STARTED:
         logger.info(_("Waiting for the state machine execution to finish"))
+        # overwriting signal handlers here does not work either
+        import rafcon
+        rafcon.core.start.register_signal_handlers(rafcon.core.start.signal_handler)
         core_singletons.state_machine_execution_engine.join()
         logger.info(_("State machine execution has finished"))
+        core_singletons.state_machine_manager.delete_all_state_machines()
 
     logger.info(_("Exiting ..."))